updates to get X11 working better with __float128 precision
Hg-commit: 9d03a2ecdca126e4d0a9be9fb5dc3ef0b669e62a

+++ src/sys/draw/utils/axis.c
@@ -93,7 +93,7 @@ PetscErrorCode PetscADefLabel(PetscReal val,PetscReal sep,char **p)
 	   use those widths */
 	if (w > 0) sprintf(fmat,"%%%d.%dlf",w + 1,d);
 	else {ierr = PetscStrcpy(fmat,"%lf");CHKERRQ(ierr);}
-	sprintf(buf,fmat,val);
+	sprintf(buf,fmat,(double)val);
         ierr = PetscStripInitialZero(buf);CHKERRQ(ierr);
         ierr = PetscStripAllZeros(buf);CHKERRQ(ierr);
         ierr = PetscStripTrailingZeros(buf);CHKERRQ(ierr);
